refactor: split UI components into framework folders
