+++ server/app/routes/admin.crawler.tsx
@@ -6,7 +6,7 @@ import {
   ReviewSource,
   sendCrawlerCommand,
 } from "~/queue-handling/review.server";
-import Button from "~/components/button";
+import Button from "~/components/tremor-ui/button";
 
 interface ActionData {
   data?: string;
